Bazi ozgarish

+++ database/migrations/2021_11_27_165031_create_food_compositions_table.php
@@ -18,7 +18,7 @@ class CreateFoodCompositionsTable extends Migration
             $table->integer('food_name_id')->references('id')->on('food');
             $table->integer('product_name_id')->references('id')->on('products');
             $table->integer('age_name_id')->references('id')->on('age_ranges');
-            $table->integer('product_weight');
+            $table->double('product_weight', 8, 3);
             $table->timestamps();
         });
     }
